requirements.txt
-e git+https://github.com/django/django.git@6a092f24e01b6448f4e5601968db82720c9397de#egg=Django-dev

gevent-socketio
jsonfield==0.9.2

django-ttag         # to support template tags in content_tags

django_compressor
-e git://github.com/DeadWisdom/pyScss.git#egg=pyscss
BeautifulSoup       # For Django Compressor